implement proper error handling

+++ api/routes/books.py
@@ -1,6 +1,6 @@
 from typing import OrderedDict
 
-from fastapi import APIRouter, status
+from fastapi import APIRouter, status, Response, HTTPException
 from fastapi.responses import JSONResponse
 
 from api.db.schemas import Book, Genre, InMemoryDB
@@ -35,6 +35,9 @@ db.books = {
 
 @router.post("/", status_code=status.HTTP_201_CREATED)
 async def create_book(book: Book):
+    if book.id in db.books:
+        raise HTTPException(status_code=400, detail="Book with this ID already exists")
+    
     db.add_book(book)
     return JSONResponse(
         status_code=status.HTTP_201_CREATED, content=book.model_dump()
@@ -47,19 +50,27 @@ async def create_book(book: Book):
 async def get_books() -> OrderedDict[int, Book]:
     return db.get_books()
 
+
 @router.get("/{book_id}", response_model=Book, status_code=status.HTTP_200_OK)
 async def get_book(book_id: int) -> Book:
-    return db.get_book(book_id)
+    book = db.get_book(book_id)
+    if book is None:
+        raise HTTPException(status_code=404, detail="Book not found")
+    return book
+
 
 @router.put("/{book_id}", response_model=Book, status_code=status.HTTP_200_OK)
 async def update_book(book_id: int, book: Book) -> Book:
-    return JSONResponse(
-        status_code=status.HTTP_200_OK,
-        content=db.update_book(book_id, book).model_dump(),
-    )
+    if book_id not in db.books:
+        raise HTTPException(status_code=404, detail="Book not found")
+
+    return db.update_book(book_id, book)
 
 
 @router.delete("/{book_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_book(book_id: int) -> None:
+async def delete_book(book_id: int):
+    if book_id not in db.books:
+        raise HTTPException(status_code=404, detail="Book not found")
+    
     db.delete_book(book_id)
-    return JSONResponse(status_code=status.HTTP_204_NO_CONTENT, content=None)
+    return Response(status_code=status.HTTP_204_NO_CONTENT)
